Action Button 快捷栏按键 / Stats Info 显示 Player 相关信息

+++ Assets/Scripts/Combat/AttackData_SO.cs
@@ -8,8 +8,8 @@ public class AttackData_SO : ScriptableObject
     public float attackRange;
     public float skillRange;
     public float coolDown;
-    public float minDamage;
-    public float maxDamage;
+    public int minDamage;
+    public int maxDamage;
     public float criticalMultiplier;
     public float criticalChance;
